requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
gunicorn==21.2.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
psycopg==3.1.12
pytest==7.4.2
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
